refactor(CarNameTest): 이름 테스트를 Cars에서 CarNameTest로 책임 이동
